[js] Organize BiDi test based on the module
